requirements.txt
# requirements.txt
pandas
psycopg2-binary
sqlalchemy
google-generativeai
python-dotenv
flask